+++ src/card/sets/y1994/legends.rs
@@ -798,6 +798,7 @@ static VISIONS_LOOK: TopCardSelectionDef = TopCardSelectionDef {
     object: None,
     minimum: 0,
     maximum: 0,
+    select_all_matching: false,
     reveal_selected: false,
     selected_zone: ZoneKind::Library,
     selected_placement: ZonePlacement::Top,