+++ src/card/sets/y1997/visions.rs
@@ -12,6 +12,7 @@ static IMPULSE_SELECTION: TopCardSelectionDef = TopCardSelectionDef {
     object: None,
     minimum: 1,
     maximum: 1,
+    select_all_matching: false,
     reveal_selected: false,
     selected_zone: ZoneKind::Hand,
     selected_placement: ZonePlacement::Top,