+++ src/card/sets/y2000/invasion.rs
@@ -54,6 +54,7 @@ static OPT_SELECTION: TopCardSelectionDef = TopCardSelectionDef {
     object: None,
     minimum: 0,
     maximum: 1,
+    select_all_matching: false,
     reveal_selected: false,
     selected_zone: ZoneKind::Library,
     selected_placement: ZonePlacement::Bottom,